Fixed DISM and Disk Cleanup
Fixed the DISM Check for both occurrences. When the EXE is run as admin it will pop up one window and run the SFC Scannow and then once that is exited it will run the DISM Checker.

Disk Cleanup was not properly configured and was not working correctly. Fixed it to work properly.